filewriter v2 docs: Update detector transformations
We removed the translation from the transformations because we decided
that the transformations group should only describe movement during
acquisition. The translation is part of the geometry group, which
describes the position of the detector in the lab.

The `depends_on` attributes of `[slow|fast]_pixel_direction` need to
refer to the detector geometry group, since they are given in detector
coordinates.